requirements.txt
asgiref==3.5.2
coverage==6.3.3
Django==4.0.4
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
djangorestframework==3.13.1
Pillow==9.1.1
djoser==2.1.0
django-cors-headers==3.12.0
djangorestframework-simplejwt==5.1.0